Fix narrowing and unused-variable warnings for MSVC and clang

+++ cudax/include/cuda/experimental/__cuco/__open_addressing/open_addressing_impl.cuh
@@ -243,7 +243,10 @@ public:
     const auto __grid_size = ::cuda::experimental::cuco::__detail::__grid_size(
       static_cast<::cuda::experimental::cuco::__detail::__index_type>(__n));
     __open_addressing::__fill<::cuda::experimental::cuco::__detail::__default_block_size()>
-      <<<__grid_size, ::cuda::experimental::cuco::__detail::__default_block_size(), 0, __stream.get()>>>(
+      <<<static_cast<unsigned int>(__grid_size),
+         ::cuda::experimental::cuco::__detail::__default_block_size(),
+         0,
+         __stream.get()>>>(
         __slots.data(), static_cast<::cuda::experimental::cuco::__detail::__index_type>(__n), __empty_slot_sentinel);
   }
 
@@ -262,7 +265,10 @@ public:
     const auto __grid_size = ::cuda::experimental::cuco::__detail::__grid_size(__num_keys, __cg_size);
 
     __open_addressing::__insert_if_n<__cg_size, ::cuda::experimental::cuco::__detail::__default_block_size()>
-      <<<__grid_size, ::cuda::experimental::cuco::__detail::__default_block_size(), 0, __stream.get()>>>(
+      <<<static_cast<unsigned int>(__grid_size),
+         ::cuda::experimental::cuco::__detail::__default_block_size(),
+         0,
+         __stream.get()>>>(
         __first,
         __num_keys,
         ::cuda::constant_iterator<bool>{true},
@@ -287,7 +293,10 @@ public:
     const auto __grid_size = ::cuda::experimental::cuco::__detail::__grid_size(__num_keys, __cg_size);
 
     __open_addressing::__insert_if_n<__cg_size, ::cuda::experimental::cuco::__detail::__default_block_size()>
-      <<<__grid_size, ::cuda::experimental::cuco::__detail::__default_block_size(), 0, __stream.get()>>>(
+      <<<static_cast<unsigned int>(__grid_size),
+         ::cuda::experimental::cuco::__detail::__default_block_size(),
+         0,
+         __stream.get()>>>(
         __first, __num_keys, ::cuda::constant_iterator<bool>{true}, ::cuda::std::identity{}, __container_ref);
   }
 
@@ -306,7 +315,10 @@ public:
     const auto __grid_size = ::cuda::experimental::cuco::__detail::__grid_size(__num_keys, __cg_size);
 
     __open_addressing::__contains_if_n<__cg_size, ::cuda::experimental::cuco::__detail::__default_block_size()>
-      <<<__grid_size, ::cuda::experimental::cuco::__detail::__default_block_size(), 0, __stream.get()>>>(
+      <<<static_cast<unsigned int>(__grid_size),
+         ::cuda::experimental::cuco::__detail::__default_block_size(),
+         0,
+         __stream.get()>>>(
         __first,
         __num_keys,
         ::cuda::constant_iterator<bool>{true},
